[#26] Remove: move TabSection, ProgressSection, UserSection to MyPage directory
- This is because TabSection, ProgressSection, MyPage are only used in MyPage
