--- src/components/MyPage/index.tsx
@@ -1,3 +1,0 @@
-export * from './ProgressSection';
-export * from './TabSection';
-export * from './UserSection';
